Scripts/Streamlit/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
snowflake
snowflake-connector-python 
openai
toml
pinecone-client